+++ Modelos/Estados.py
@@ -1,6 +1,5 @@
-from datetime import datetime
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String, Float, ForeignKey, Date
+from sqlalchemy import Column, Integer, String
 from Modelos.Database.Restaurant import Restaurant 
 
 #----------- DAO -----------------
@@ -16,10 +15,4 @@ class Estado(Base):
     def __str__(self):
         return  "ID: " + self.id_estado 
 
-Base.metadata.create_all(Restaurant.getInstance().getEngine())
-#----------- SERVICIO -----------------
-
-def getEstadoById(id_estado):
-    session = Restaurant.getInstance().session
-    estado = session.query(Estado).filter_by(id_estado=id_estado).one()
-    return estado
+Base.metadata.create_all(Restaurant.getInstance().engine)
