+++ Modelos/Productos.py
@@ -1,11 +1,7 @@
-from datetime import datetime
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy import Column, Integer, String, Float
-from sqlalchemy.orm import sessionmaker, session
-
 from Modelos.Database.Restaurant import Restaurant 
 
-#----------- DAO -----------------
 Base = declarative_base()
 
 class Producto(Base):
@@ -25,25 +21,8 @@ class Producto(Base):
     def __str__(self):
         return  "ID: " + str(self.id_producto) + " NOMBRE: " + self.nombre_producto
 
-Base.metadata.create_all(Restaurant.getInstance().getEngine())
-    
-#--------- SERVICIOS ------------ 
+Base.metadata.create_all(Restaurant.getInstance().engine)
 
-def getAllProductos():
-    session = Restaurant.getInstance().session
-    productos = session.query(Producto).all()
-    return productos
-
-def getProductoById(id:int):
-    session = Restaurant.getInstance().session
-    try:
-        producto = session.query(Producto).filter_by(id_producto = id).one()
-        return producto
-    except Exception as e:
-        print(e)
-        return None
-    
-    
 """
 session = Restaurant.getInstance().session
 productos = session.query(Producto).all()
